[svelte] Rename TT api file
